refactoring : make showcase become component
